@@ -74,10 +125,16 @@
     "husky": "^9.0.11",
     "jest": "^29.7.0",
     "lint-staged": "^15.2.2",
+    "nock": "^14.0.5",
     "prettier": "^3.2.5",
+    "semantic-release": "^22.0.12",
+    "size-limit": "^11.2.0",
+    "supertest": "^7.1.3",
     "syncpack": "^12.3.0",
+    "ts-jest": "^29.4.0",
+    "tsup": "^8.5.0",
     "turbo": "^1.13.0",
-    "typescript": "^5.4.3"
+    "typescript": "^5.6.3"
   },
   "engines": {
     "node": ">=18.0.0",
